bug found
bug found and fixed. the problem was the game level wasnt A) accessing the correct spawn point in the inspector. and B) the game level was trying to acess the static instance before it was intialised. problem fixed by moveing initaliaztion to the awake method and the instance get to the start method