scene loads in from a start button
the scene now loads in from the start button. It follows a pattern of the main game objects (player, cam etc) are loaded into a main scene and the level objects are loaded into a second scene and a third scene is created to hold the object pool to keep track of everything without clogging up the main scenes. better for object management